Restructure to put type definitions in better places
